% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{api1_url}
\alias{api1_url}
\title{Get partial API v1 URL for the provided path}
\usage{
api1_url(path)
}
\arguments{
\item{path}{relative path of the API v1 endpoint (e.g. \code{/me})}
}
\value{
Returns "/api/v1" + path
}
\description{
Get partial API v1 URL for the provided path
}
\examples{
 api1_url("/me")
}
