% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{api2_url}
\alias{api2_url}
\title{Get partial API v2 URL for the provided path}
\usage{
api2_url(path)
}
\arguments{
\item{path}{relative path of the API v2 endpoint (e.g. \code{/me})}
}
\value{
Returns "/api/v2" + path
}
\description{
Get partial API v2 URL for the provided path
}
\examples{
 api2_url("/me")
}
