% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{authorization}
\alias{authorization}
\title{Provide OAuth2 token to use for authorization with Genesys}
\usage{
authorization(authorization)
}
\arguments{
\item{authorization}{OAuth2 Authorization header obtained from somewhere else (e.g. an ENV variable)}
}
\description{
Provide OAuth2 token to use for authorization with Genesys
}
\seealso{
\code{\link{user_login}}, \code{\link{client_login}}
}
