% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{check_country}
\alias{check_country}
\title{Run Land-or-Sea check on MCPD data. Uploads only rows where ORIGCTY,
DECLATITUDE and DECLONGITUDE are provided.}
\usage{
check_country(mcpd)
}
\arguments{
\item{mcpd}{Accession passport data in MCPD format}
}
\value{
Results from validator
}
\description{
Run Land-or-Sea check on MCPD data. Uploads only rows where ORIGCTY,
DECLATITUDE and DECLONGITUDE are provided.
}
\examples{
\dontrun{
  geoCheck <- genesysr::check_country(mcpd)
}

}
