% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{check_landorsea}
\alias{check_landorsea}
\title{Run Land-or-Sea check on MCPD data using https://validator.genesys-pgr.org.
Uploads only rows where DECLATITUDE and DECLONGITUDE are provided. 
In practice it is better to use `check_country` if ORIGCTY data exists.}
\usage{
check_landorsea(mcpd)
}
\arguments{
\item{mcpd}{Accession passport data in MCPD format}
}
\value{
Results from validator
}
\description{
Run Land-or-Sea check on MCPD data using https://validator.genesys-pgr.org.
Uploads only rows where DECLATITUDE and DECLONGITUDE are provided. 
In practice it is better to use `check_country` if ORIGCTY data exists.
}
\examples{
\dontrun{
  waterCheck <- genesysr::check_landorsea(mcpd)
}

}
