% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{check_taxonomy}
\alias{check_taxonomy}
\title{Check MCPD taxonomic data (GENUS, SPECIES, SPAUTHOR, SUBTAXA, SUBTAUTHOR)
using https://validator.genesys-pgr.org.}
\usage{
check_taxonomy(mcpd, toCurrentTaxa = FALSE)
}
\arguments{
\item{mcpd}{Accession passport data in MCPD format}

\item{toCurrentTaxa}{Should obsoleted names be reported?}
}
\value{
Results from validator
}
\description{
Duplicate input rows are removed using dplyr::distinct() and results are
returned for unique rows.
}
\examples{
\dontrun{
  taxaCheck <- genesysr::check_taxonomy(mcpd)
}

}
