% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{client_login}
\alias{client_login}
\title{Login to Genesys as a service client (system-to-system)}
\usage{
client_login()
}
\description{
The client must be enabled for Client Credential grant on Genesys.
}
\seealso{
\code{\link{setup}}
}
