% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\name{.fetch_csv_page}
\alias{.fetch_csv_page}
\title{Fetch a page of data from Genesys}
\usage{
.fetch_csv_page(
  path,
  filters = list(),
  accept = "text/csv",
  page = 0,
  size = 1000
)
}
\arguments{
\item{path}{API path}

\item{filters}{Filters}

\item{accept}{Accepted content type}

\item{page}{Page to request}

\item{size}{Size of page}
}
\description{
Fetch a page of data from Genesys
}
\keyword{internal}
