% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\name{.fieldsToSelect}
\alias{.fieldsToSelect}
\title{Map MCPD column names to their Genesys JSON equivalents}
\usage{
.fieldsToSelect(fields = NULL)
}
\description{
Map MCPD column names to their Genesys JSON equivalents
}
\keyword{internal}
