% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\name{.list_accessions_page}
\alias{.list_accessions_page}
\title{Get one page of accession passport data}
\usage{
.list_accessions_page(filters, page = 0, size = 1e+06, select = NULL)
}
\value{
table
}
\description{
Get one page of accession passport data
}
\keyword{internal}
