% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\docType{data}
\name{.MAX_ALLOWED_PAGES}
\alias{.MAX_ALLOWED_PAGES}
\title{Max pages to retrieve}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.MAX_ALLOWED_PAGES
}
\description{
Max pages to retrieve
}
\keyword{internal}
