% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\name{download_pdci}
\alias{download_pdci}
\title{Download PDCI data for one genebank in Excel format and save it to disk.}
\usage{
download_pdci(instituteCode, file = NULL)
}
\arguments{
\item{instituteCode}{FAO WIEWS institute code}

\item{file}{Target file name. Defaults to Genesys-provided file name in the current working directory.}
}
\value{
The downloaded PDCI file name
}
\description{
Download PDCI data for one genebank in Excel format and save it to disk.
}
\examples{
\dontrun{
  # Download PDCI  data for NGA039
  excelData <- download_pdci("NGA039")
}

}
