% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_DOI}
\alias{filter_DOI}
\title{Add filter on accession DOI}
\usage{
filter_DOI(filter = list(), DOI)
}
\arguments{
\item{filter}{Existing filters (or blank list if not provided)}

\item{DOI}{Accession DOI}
}
\description{
Add filter on accession DOI
}
