% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_GENUS}
\alias{filter_GENUS}
\title{Add filter by genus}
\usage{
filter_GENUS(filter = list(), GENUS)
}
\arguments{
\item{filter}{Existing filters (or blank list if not provided)}

\item{GENUS}{List of genera}
}
\description{
Add filter by genus
}
