% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_INSTCODE}
\alias{filter_INSTCODE}
\title{Add filter by genus}
\usage{
filter_INSTCODE(filter = list(), INSTCODE)
}
\arguments{
\item{filter}{Existing filters (or blank list if not provided)}

\item{INSTCODE}{List of WIEWS institute codes}
}
\description{
Add filter by genus
}
