% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_ORIGCTY}
\alias{filter_ORIGCTY}
\title{Add filter on Country of origin of material}
\usage{
filter_ORIGCTY(filter = list(), ORIGCTY)
}
\arguments{
\item{filter}{Existing filters (or blank list if not provided)}

\item{ORIGCTY}{Country of origin}
}
\description{
Add filter on Country of origin of material
}
