% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_SAMPSTAT}
\alias{filter_SAMPSTAT}
\title{Add filter on Biological status of sample}
\usage{
filter_SAMPSTAT(filter = list(), SAMPSTAT)
}
\arguments{
\item{filter}{Existing filters (or blank list if not provided)}

\item{SAMPSTAT}{Biological status of sample}
}
\description{
Add filter on Biological status of sample
}
