% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_SPECIES}
\alias{filter_SPECIES}
\title{Add filter on specific epithet}
\usage{
filter_SPECIES(filter = list(), SPECIES)
}
\arguments{
\item{filter}{Existing filters (or blank list if not provided)}

\item{SPECIES}{List of specific epithets}
}
\description{
Add filter on specific epithet
}
