% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\name{list_crops}
\alias{list_crops}
\title{Fetch Genesys crops. Note that the list of Genesys crops does not fully
correspond with various CROPNAME in MCPD provided by genebanks.}
\usage{
list_crops()
}
\value{
Genesys crops
}
\description{
Fetch Genesys crops. Note that the list of Genesys crops does not fully
correspond with various CROPNAME in MCPD provided by genebanks.
}
\examples{
\dontrun{
  # Retrieve all Genesys crops
  crops <- genesysr::list_crops()
}

}
