% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\name{list_species}
\alias{list_species}
\title{Fetch taxonomic data of selected accessions.}
\usage{
list_species(filters = list())
}
\arguments{
\item{filters}{an R \code{structure} with Genesys filters}
}
\value{
Taxonomic records of selected accessions
}
\description{
Fetch taxonomic data of selected accessions.
}
\examples{
\dontrun{
  # Retrieve taxa of selected accessions
  taxa <- genesysr::list_species(mcpd_filter(INSTCODE = c("LBN002", "MEX002")))
}

}
\seealso{
\code{\link{mcpd_filter}}
}
