% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\name{me}
\alias{me}
\title{Who am i? Loads and prints the user profile from Genesys as JSON.
You need to be logged in.}
\usage{
me()
}
\description{
Who am i? Loads and prints the user profile from Genesys as JSON.
You need to be logged in.
}
\examples{
\dontrun{
  # Login
  setup_production()
  user_login()
  me()
}

}
