% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\docType{data}
\name{SELECT_ALL_FIELDS}
\alias{SELECT_ALL_FIELDS}
\title{List of all supported passport data fields}
\format{
An object of class \code{character} of length 49.
}
\usage{
SELECT_ALL_FIELDS
}
\description{
List of all supported passport data fields
}
\keyword{datasets}
