% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\docType{data}
\name{SELECT_GEO_FIELDS}
\alias{SELECT_GEO_FIELDS}
\title{List of passport data fields relevant for GIS analysis}
\format{
An object of class \code{character} of length 16.
}
\usage{
SELECT_GEO_FIELDS
}
\description{
List of passport data fields relevant for GIS analysis
}
\keyword{datasets}
