% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\docType{data}
\name{SELECT_INFO_FIELDS}
\alias{SELECT_INFO_FIELDS}
\title{List of basic passport data fields}
\format{
An object of class \code{character} of length 8.
}
\usage{
SELECT_INFO_FIELDS
}
\description{
List of basic passport data fields
}
\keyword{datasets}
