% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{setup}
\alias{setup}
\title{Configure the Genesys environment}
\usage{
setup(server = NULL, client_id = NULL, client_secret = NULL)
}
\arguments{
\item{server}{Server base URL (e.g. "https://api.genesys-pgr.org" or "https://api.sandbox.genesys-pgr.org")}

\item{client_id}{OAuth client ID}

\item{client_secret}{OAuth client secret}
}
\description{
Configure the Genesys environment
}
\examples{
  # Link with sandbox
  setup_sandbox()

}
\seealso{
See utility methods \code{\link{setup_production}}, \code{\link{setup_sandbox}}
}
