% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{setup_production}
\alias{setup_production}
\title{Setup for Genesys Production}
\usage{
setup_production()
}
\description{
Use the Genesys R Client with <https://api.genesys-pgr.org> requiring \code{\link{user_login}}
}
