% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{setup_sandbox}
\alias{setup_sandbox}
\title{Setup for Genesys Sandbox}
\usage{
setup_sandbox()
}
\description{
Use the Genesys R Client with <https://api.sandbox.genesys-pgr.org> requiring \code{\link{user_login}}
}
