% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{.postForm}
\alias{.postForm}
\title{HTTP POST method}
\usage{
.postForm(path, query = NULL, accept = "application/json", ...)
}
\arguments{
\item{path}{API path}

\item{query}{query string parameters}

\item{accept}{Content type to use in Accept header}

\item{...}{parameters to submit in form body}
}
\value{
httr response
}
\description{
HTTP POST method
}
\keyword{internal}
